simplify and reorganize code generation a tad.
Tested-on: collins (Linux 2.6.32-5-amd64)

+++ secretrun/secretrun.go
@@ -4,10 +4,12 @@ import (
 	"fmt"
 	"os"
 	"syscall"
-	"github.com/droundy/goadmin/ago"
+	"github.com/droundy/goadmin/ago/compile"
 	"github.com/droundy/goadmin/deps"
 	"github.com/droundy/goopt"
 )
+
+var outname = goopt.String([]string{"-o"}, "FILENAME.go", "the name of the generated go file")
 
 func main() {
 	goopt.Parse(func() []string { return []string{} })
@@ -20,12 +22,16 @@ func main() {
 	}
 
 	execname := goopt.Args[0]+".secret"
-	e := ago.Compile(execname, goopt.Args)
+	e := compile.Compile(execname, goopt.Args)
 	if e != nil {
 		fmt.Fprintln(os.Stderr, e)
 		os.Exit(1)
 	}
-	e = deps.Exec("./"+execname)
+	if *outname == "FILENAME.go" {
+		fmt.Fprintln(os.Stderr, "secretrun requires a -o argument!")
+		os.Exit(1)
+	}
+	e = deps.Exec("./"+execname, *outname)
 	os.Remove(execname)
 	if e != nil {
 		fmt.Fprintln(os.Stderr, e)
